Move primo to utils
